Cleaning up Impact View Controller
In this commit, I used the Gravity Slider Flow Layout to illustrate the amount of gallons of water used per week and year. I also show how much water users can save if they decrease their shower time by one tenth of their average shower time. The purpose of this is to show that even saving a few seconds of water can make a huge difference in the long run.