Adding the circular timer feature for routines
In this commit, I used the CircularSlider library to create a circular timer for each step of the routine. This is shown in RoutineTimerViewController, where each cell of the collection view is a step of the routine. After the time is up for a particular step, the collection view programatically scrolls to the next cell.