Making sure the username shows up right after making an account
Originally, if the new user signed up and logged in immediately, their username was not showing up on the home page. This is because registering a new user in parse happens asynchronously. I added a completion block to the register method to go the next screen only after the registering is completed